code/scripts/requirements.txt
pycocotools
pyyaml
pysqlite3
